backend/youtube-summariser/requirements.txt
fastapi
uvicorn
youtube-transcript-api
langchain
python-dotenv
groq